--- cwms-radar-client/src/test/java/mil/army/usace/hec/cwms/radar/client/TestController.java
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) 2021
- * United States Army Corps of Engineers - Hydrologic Engineering Center (USACE/HEC)
- * All Rights Reserved.  USACE PROPRIETARY/CONFIDENTIAL.
- * Source may not be released without written approval from HEC
- */
-
-package mil.army.usace.hec.cwms.radar.client;
-
-import java.io.File;
-import java.io.IOException;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import okhttp3.mockwebserver.MockWebServer;
-
-abstract class TestController {
-    final String readJsonFile(MockWebServer mockWebServer, String jsonPath) throws IOException {
-        URL resource = getClass().getClassLoader().getResource(jsonPath);
-        if (resource == null) {
-            throw new IOException("Resource not found: " + jsonPath);
-        }
-        Path path = new File(resource.getFile()).toPath();
-        return String.join("\n", Files.readAllLines(path));
-    }
-}
